ref(crashContent): Move crash content related files to folder
