+++ static/app/components/events/interfaces/crashContent/exception/index.tsx
@@ -1,9 +1,10 @@
 import ErrorBoundary from 'app/components/errorBoundary';
-import ExceptionContent from 'app/components/events/interfaces/exceptionContent';
 import RawExceptionContent from 'app/components/events/interfaces/rawExceptionContent';
 import {ExceptionType, Group, PlatformType, Project} from 'app/types';
 import {Event} from 'app/types/event';
 import {STACK_TYPE, STACK_VIEW} from 'app/types/stacktrace';
+
+import Content from './content';
 
 type Props = {
   stackType: STACK_TYPE;
@@ -37,7 +38,7 @@ const Exception = ({
         platform={platform}
       />
     ) : (
-      <ExceptionContent
+      <Content
         type={stackType}
         stackView={stackView}
         values={values}
